\docType{package}
\name{ksplines}
\alias{ksplines}
\alias{ksplines-package}
\title{Fit a spline mixture to curves}
\usage{
ksplines(curves, k, df, order, lambda, xrange = NULL, tol = 0.01,
  verbose = FALSE, seed = 1)
}
\description{
Fit a spline mixture to curves

Functions for clustering time series data using splines.
}

