\docType{package}
\name{ksplines}
\alias{ksplines}
\alias{ksplines-package}
\title{Learn spline codebook for sparse curves}
\usage{
ksplines(curves, k, df, degree, intercept = TRUE, lambda = 0.1,
  maxiter = 10)
}
\description{
Learn spline codebook for sparse curves

Functions for clustering time series data using splines.
}

