\docType{package}
\name{ksplines}
\alias{ksplines}
\alias{ksplines-package}
\title{Learn spline codebook for sparse curves}
\usage{
ksplines(curves, k, df, order, lambda = 0.1, maxiter = 10,
  normalize = FALSE)
}
\description{
Learn spline codebook for sparse curves

Functions for clustering time series data using splines.
}

